feat: add string array tests in main in ex02

+++ CPP07/ex01/iter.hpp
@@ -14,9 +14,9 @@
 #define ITER_HPP
 
 template <typename T>
-void	iter(T *array, int length, void (*f)(T &))
+void	iter(T *array, size_t length, void (*f)(T &))
 {
-	for (int i = 0; i < length; i++)
+	for (size_t i = 0; i < length; i++)
 		f(array[i]);
 }
